+++ internal/clioptions/clioptions.go
@@ -46,12 +46,17 @@ type CLIOptions struct {
 	DeployType string
 	NoSemVer   bool
 
+	IAMRole string
+
+	UserEmail string
+
+	UserEmails []string
+	GroupID    string
+
 	BasicClientID     string
 	BasicClientSecret string
 	JWTJsonPath       string
-
-	ServiceAccountRole string
-	OutputPath         string
+	OutputPath        string
 
 	MarketplaceResourcePaths []string
 	// MarketplaceItemID is the itemId field of a Marketplace item
@@ -132,12 +137,26 @@ func (o *CLIOptions) AddContextAuthFlags(flags *pflag.FlagSet) {
 }
 
 func (o *CLIOptions) AddServiceAccountFlags(flags *pflag.FlagSet) {
-	flags.StringVarP(&o.ServiceAccountRole, "role", "r", "", "the company role of the service account")
+	flags.StringVarP(&o.IAMRole, "role", "r", "", "the company role of the service account")
 }
 
 func (o *CLIOptions) AddJWTServiceAccountFlags(flags *pflag.FlagSet) {
 	o.AddServiceAccountFlags(flags)
-	flags.StringVarP(&o.OutputPath, "output", "o", "", "write the service account to a file")
+	flags.StringVarP(&o.OutputPath, "output", "o", "", "write the service account configuration as json to a file")
+}
+
+func (o *CLIOptions) AddNewUserFlags(flags *pflag.FlagSet) {
+	flags.StringVarP(&o.IAMRole, "role", "r", "", "the company role of the user")
+	flags.StringVarP(&o.UserEmail, "email", "", "", "the email of the user to add")
+}
+
+func (o *CLIOptions) CreateNewGroupFlags(flags *pflag.FlagSet) {
+	flags.StringVarP(&o.IAMRole, "role", "r", "", "the company role of the group")
+}
+
+func (o *CLIOptions) AddMemberToGroupFlags(flags *pflag.FlagSet) {
+	flags.StringSliceVarP(&o.UserEmails, "user-email", "", []string{}, "the list of user email to add to the group")
+	flags.StringVarP(&o.GroupID, "group-id", "", "", "the group id where to add the users")
 }
 
 func (o *CLIOptions) AddMarketplaceApplyFlags(cmd *cobra.Command) {